add checks to for comment owner before modifying/deleting

+++ api/app/Http/Controllers/CommentController.php
@@ -60,6 +60,7 @@ class CommentController extends Controller
      */
     public function update(Request $request, Comment $comment)
     {
+        abort_if($comment->user_id !== auth()->user()->id, 500, 'You are not authorized to do this.');
         // dd($request->input(), $comment);
         $comment->fill($request->only('content'));
         $comment->save();
@@ -71,6 +72,9 @@ class CommentController extends Controller
      */
     public function destroy(Comment $comment)
     {
+        abort_if($comment->user_id !== auth()->user()->id, 500, 'You are not authorized to do this.');
+        
+        $comment->replies()->delete();
         $comment->delete();
         return response()->json(null);
     }
